from FBV to CBV

+++ UserApp/models.py
@@ -13,7 +13,7 @@ class Customer(models.Model):
 class Product(models.Model):
     #one product
     name = models.CharField(max_length=200)
-    price = models.FloatField()
+    price = models.DecimalField(max_digits=8, decimal_places=2)
     digital = models.BooleanField(default=False)
     product_Img=models.ImageField(upload_to='productsImg', null=True)
